Improve subclassability/extendability of attribute types.
Attribute now use class-scope defaults for most, if not all, Attribute
properties. That allows subclasses to have a common setup, e.g. an Email
type that is a String subclass that always provides a LooksLikeAnEmail
validator.

Purely as a nod to syntax, the default validator is an Always validator.
That allows:

SomeType(validator=All(SomeType.validator, MyValidator()))

Instead of:

SomeType(validator=(MyValidator() if not SomeType.validator else
All(SomeType.validator, MyValidator())))

The 2nd version is very ugly and uses duplicate code.