Fix Sequence to re-raise exceptions with correctly mapped names.
When the Invalid exception from a child attribute was sucked into the
sequence's error_dict the key name was being incorrectly generated for
errors from immediate child attributes, i.e. those errors that have a
key of ''.

This is basically the same change that had already been applied in
Structure.validate().